Nhanh cua anh ngoc

+++ src/main/java/com/example/shopgiayonepoly/controller/CustomerController.java
@@ -50,7 +50,14 @@ public class CustomerController {
 //    }
 
     @GetMapping("/list")
-    public String getListCustomrByPage(@RequestParam(name = "pageNumber", defaultValue = "0") Integer pageNumber, Model model) {
+    public String getListCustomrByPage(@RequestParam(name = "pageNumber", defaultValue = "0") Integer pageNumber, Model model, HttpSession session) {
+        Staff staffLogin = (Staff) session.getAttribute("staffLogin");
+        if (staffLogin == null) {
+            return "redirect:/login";
+        }
+        if(staffLogin.getStatus() != 1) {
+            return "redirect:/home_manage";
+        }
         Pageable pageable = PageRequest.of(pageNumber, pageSize);
         Page<CustomerResponse> pageCustomer = customerService.getAllCustomerByPage(pageable);
         model.addAttribute("pageCustomer", pageCustomer);
@@ -80,7 +87,14 @@ public class CustomerController {
     }
 
     @PostMapping("/add")
-    public String addCustomer(Model model, @Valid @ModelAttribute(name = "customer") CustomerRequest customerRequest, BindingResult result) throws IOException {
+    public String addCustomer(Model model, @Valid @ModelAttribute(name = "customer") CustomerRequest customerRequest, BindingResult result, HttpSession session) throws IOException {
+        Staff staffLogin = (Staff) session.getAttribute("staffLogin");
+        if (staffLogin == null) {
+            return "redirect:/login";
+        }
+        if(staffLogin.getStatus() != 1) {
+            return "redirect:/home_manage";
+        }
         // Kiểm tra tên
         if (customerRequest.getFullName() == null || customerRequest.getFullName().trim().isEmpty()) {
             result.rejectValue("fullName", "error.customer", "Tên không được để trống!"); // Thông báo nếu tên rỗng hoặc chỉ chứa khoảng trắng
@@ -151,7 +165,14 @@ public class CustomerController {
     }
 
     @GetMapping("/edit/{id}")
-    public String editCustomer(Model model, @PathVariable("id") Integer id) {
+    public String editCustomer(Model model, @PathVariable("id") Integer id, HttpSession session) {
+        Staff staffLogin = (Staff) session.getAttribute("staffLogin");
+        if (staffLogin == null) {
+            return "redirect:/login";
+        }
+        if(staffLogin.getStatus() != 1) {
+            return "redirect:/home_manage";
+        }
         Customer customer = customerService.getOne(id);
         CustomerRequest customerRequest = new CustomerRequest();
         customerRequest.setId(customer.getId());
@@ -173,7 +194,14 @@ public class CustomerController {
     }
 
     @PostMapping("/update")
-    public String updateCustomer(Model model, @Valid @ModelAttribute(name = "customer") CustomerRequest customerRequest, BindingResult result) throws IOException {
+    public String updateCustomer(Model model, @Valid @ModelAttribute(name = "customer") CustomerRequest customerRequest, BindingResult result, HttpSession session) throws IOException {
+        Staff staffLogin = (Staff) session.getAttribute("staffLogin");
+        if (staffLogin == null) {
+            return "redirect:/login";
+        }
+        if(staffLogin.getStatus() != 1) {
+            return "redirect:/home_manage";
+        }
         // Kiểm tra tên
         if (customerRequest.getFullName() == null || customerRequest.getFullName().trim().isEmpty()) {
             result.rejectValue("fullName", "error.customer", "Tên không được để trống!"); // Thông báo nếu tên rỗng hoặc chỉ chứa khoảng trắng
@@ -210,8 +238,7 @@ public class CustomerController {
         }
         if (result.hasErrors()) {
             model.addAttribute("mes", "Thêm thất bại");
-            // Nếu có lỗi, trả về trang form để người dùng sửa lại
-            return "Customer/update"; // Bạn có thể trả về tên view của form nhập liệu
+            return "Customer/update";
         }
         Customer customer = customerService.getCustomerByID(customerRequest.getId());
         customerRequest.setCreateDate(customer.getCreateDate());
@@ -240,7 +267,14 @@ public class CustomerController {
     }
 
     @GetMapping("/detail/{id}")
-    public String detailCustomer(Model model, @PathVariable("id") Integer id) {
+    public String detailCustomer(Model model, @PathVariable("id") Integer id, HttpSession session) {
+        Staff staffLogin = (Staff) session.getAttribute("staffLogin");
+        if (staffLogin == null) {
+            return "redirect:/login";
+        }
+        if(staffLogin.getStatus() != 1) {
+            return "redirect:/home_manage";
+        }
         Customer customer = customerService.getOne(id);
         CustomerRequest customerRequest = new CustomerRequest();
         customerRequest.setId(customer.getId());
@@ -262,7 +296,14 @@ public class CustomerController {
     }
 
     @GetMapping("/delete/{id}")
-    public String deleteCustomer(RedirectAttributes ra, @PathVariable("id") Integer id) {
+    public String deleteCustomer(RedirectAttributes ra, @PathVariable("id") Integer id, HttpSession session) {
+        Staff staffLogin = (Staff) session.getAttribute("staffLogin");
+        if (staffLogin == null) {
+            return "redirect:/login";
+        }
+        if(staffLogin.getStatus() != 1) {
+            return "redirect:/home_manage";
+        }
         customerService.deleteCustomer(id);
         ra.addFlashAttribute("mes", "Xóa thành công Khach hang với ID là: " + id);
         return "redirect:/customer/list";
